Support Strings - and `SELECT str, sum(x) from tbl group by str` (#255)
- Fix VarChar transfer issue
- Introduce `VeScalarType` wihch is a subset of `VeType`. the other variation is `VeString`/`CVarChar`. Strings cannot be represented as Scalars and so need to be _entirely differently_ in the C code.
- Implement strings for Projection, Filter and GroupBy in the framework
- Implement strings for GroupBy in `VeRewriteStrategy`
- Create a concept of `StringProducer`. This represents an imperative approach to . The optimized versions will come from the C experts - the focus is _entirely_ on getting the framework and correctness right. `StringProducer` in this case produces data to a specific VE output. It represents the creation of a single String, and this is essential in order for us to support the widest functionality possible from Spark queries. Later, we will map these `StringProducer`s to Frovedis functions or something more parallel.
- Introduced `StringGroupByTesting` to verify it works indeed. All the things previously were derived from work on `SubstringTesting` (eg to be able to support dynamic String operations), which led to `StringProducer`. It works fine on VE.

+++ src/main/scala/com/nec/arrow/ArrowInterfaces.scala
@@ -89,7 +89,8 @@ object ArrowInterfaces {
   def c_nullable_bigint_vector(bigIntVector: BigIntVector): nullable_bigint_vector = {
     val vc = new nullable_bigint_vector()
     vc.data = bigIntVector.getDataBuffer.nioBuffer().asInstanceOf[DirectBuffer].address()
-    vc.validityBuffer = bigIntVector.getValidityBuffer.nioBuffer().asInstanceOf[DirectBuffer].address()
+    vc.validityBuffer =
+      bigIntVector.getValidityBuffer.nioBuffer().asInstanceOf[DirectBuffer].address()
     vc.count = bigIntVector.getValueCount
     vc
   }
@@ -189,19 +190,15 @@ object ArrowInterfaces {
     input: nullable_varchar_vector,
     varCharVector: VarCharVector
   ): Unit = {
-    varCharVector.allocateNew(input.size(), input.count)
+    varCharVector.allocateNew(input.size.toLong, input.count)
     varCharVector.setValueCount(input.count)
     getUnsafe.copyMemory(
       input.validityBuffer,
       varCharVector.getValidityBufferAddress,
       Math.ceil(input.count / 8.0).toInt
     )
-    getUnsafe.copyMemory(input.data, varCharVector.getDataBufferAddress, input.size())
-    getUnsafe.copyMemory(
-      input.offsets,
-      varCharVector.getOffsetBufferAddress,
-      4 * (input.count + 1)
-    )
+    getUnsafe.copyMemory(input.data, varCharVector.getDataBufferAddress, input.size.toLong)
+    getUnsafe.copyMemory(input.offsets, varCharVector.getOffsetBufferAddress, 4 * (input.count + 1))
   }
 
 }
